🤖 refactor: confirmDialog -\> Modal, DialogBase로 공통 로직, ui 분
