+++ src/components/global/Dialog/DialogBase.css.ts
@@ -37,12 +37,6 @@ export const modalConatainer = recipe({
     maxHeight: 'calc(85vh)'
   },
   variants: {
-    variant: {
-      basic: {
-        padding: '24px 24px 20px',
-        width: '300px'
-      }
-    },
     size: {
       sm: {
         transform: 'translate(-50%, -100%)'
@@ -52,35 +46,4 @@ export const modalConatainer = recipe({
       }
     }
   }
-});
-
-export const header = style({
-  marginBottom: 8,
-  zIndex: 1,
-  position: 'relative',
-  height: 24
-});
-
-export const closeIcon = style({
-  cursor: 'pointer',
-  position: 'absolute',
-  right: 0
-});
-
-export const childrenWarp = style({
-  width: '100%',
-  height: 'auto'
-});
-
-export const contents = style({
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'center',
-  height: '100%',
-  width: '100%'
-});
-
-export const buttonWrapper = style({
-  display: 'flex',
-  gap: 10
 });
